+++ frontend/src/scenes/session-recordings/player/player-meta/playerMetaLogic.test.ts
@@ -2,7 +2,7 @@ import { expectLogic } from 'kea-test-utils'
 
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { playerMetaLogic } from 'scenes/session-recordings/player/player-meta/playerMetaLogic'
-import { sessionRecordingDataLogic } from 'scenes/session-recordings/player/sessionRecordingDataLogic'
+import { sessionRecordingDataCoordinatorLogic } from 'scenes/session-recordings/player/sessionRecordingDataCoordinatorLogic'
 import { sessionRecordingPlayerLogic } from 'scenes/session-recordings/player/sessionRecordingPlayerLogic'
 
 import { useMocks } from '~/mocks/jest'
@@ -38,7 +38,7 @@ describe('playerMetaLogic', () => {
     describe('core assumptions', () => {
         it('mounts other logics', () => {
             expectLogic(logic).toMount([
-                sessionRecordingDataLogic(playerProps),
+                sessionRecordingDataCoordinatorLogic(playerProps),
                 sessionRecordingPlayerLogic(playerProps),
             ])
         })
@@ -55,7 +55,7 @@ describe('playerMetaLogic', () => {
                 id: '1',
             } as SessionRecordingType
             await expectLogic(logic, () => {
-                sessionRecordingDataLogic(playerProps).actions.loadRecordingMeta()
+                sessionRecordingDataCoordinatorLogic(playerProps).actions.loadRecordingMeta()
                 logic.actions.maybeLoadPropertiesForSessions([session])
             })
                 .toDispatchActions(['loadRecordingMetaSuccess', 'loadPropertiesForSessionsSuccess'])
